System Test - Updated ADC databus pin constraints
Noticed tons of noise when I plugged in ADC databus to FPGA. Before,
attributed this to pre-amp, and solved with 10nF caps to gnd off ADC
inputs. This time that failed, but we're more than meeting the
level-shifter's timing requirements and slew rate requirements, so why
is the level shifter outputting traces that barely reach high/low before
the next transition? Suspect this as the source of noise; realized I
forgot to make ADC databus byte high-slew rate in the constraints during
design implementation.