System Test - Added Synchronization on ADC Databus
Didn't think it was necessary, given that I deal with the ADC as a
peripheral and have a ton of wait states, but it's possible that it
routed as having the pad directly driving FPGA internals (?). Seeing a
lot of noise when I connect logic-level converter outputs to FPGA PMODs.